requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.4.1=pypi_0
blas=1.0=mkl
brotli-python=1.1.0=py310hc6cd4ac_1
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pyhd8ed1ab_0
freetype=2.12.1=h267a509_2
gmp=6.3.0=h59595ed_1
gmpy2=2.1.2=py310h3ec546c_1
gnutls=3.6.13=h85f3911_1
icu=73.2=h59595ed_0
idna=3.6=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pypi_0
ipython=8.22.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
jupyter-client=8.6.1=pypi_0
jupyter-core=5.7.2=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h0b41bf4_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=13.2.0=h807b86a_5
libhwloc=2.9.3=default_h554bfaf_1009
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=5_h92ddd45_netlib
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.43=h2797004_0
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.6=h232c23b_1
libzlib=1.2.13=hd590300_5
llvm-openmp=15.0.7=h0cdce71_0
markupsafe=2.1.5=py310h2372a71_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h213fc3f_46344
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_0
mpmath=1.3.0=pyhd8ed1ab_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pypi_0
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.4=py310hb13e2d6_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.2.1=hd590300_1
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pillow=9.4.0=py310h023d228_1
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pygments=2.17.2=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.14=hd12c33a_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=4_cp310
pytorch=2.2.1=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h2372a71_1
pyzmq=25.1.2=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.4.1.post1=py310h1fdf081_0
scipy=1.12.0=py310hb13e2d6_2
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.11.0=h00ab1b0_1
threadpoolctl=3.4.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=2.2.1=py310_cu121
torchtriton=2.2.0=py310
torchvision=0.17.1=py310_cu121
tornado=6.4=pypi_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.14.2=pypi_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
